+++ handlers/eventHandler.go
@@ -9,7 +9,7 @@ import (
 var twoYearsSellLink = "https://fr.mos.ru/pokupka-nedvizhimosti-dlya-vseh/ajax.php?category[]=PARTICIPANTS&y2_sell=1&status[]=PROCESSING&status[]=FINISHED&price_min=0&price_max=100000000000000&price_m_min=0&price_m_max=100000000000&area_min=0&area_max=100000000&floor_min=-1&floor_max=100000000&open_sale=0&pagesize=100000"
 var inMomentSellLink = "https://fr.mos.ru/pokupka-nedvizhimosti-dlya-vseh/ajax.php?category[]=PARTICIPANTS&for_sell=1&status[]=PROCESSING&status[]=FINISHED&price_min=0&price_max=100000000000000&price_m_min=0&price_m_max=100000000000&area_min=0&area_max=100000000&floor_min=-1&floor_max=100000000&open_sale=0&pagesize=100000"
 
-func (h *TrackingHandler) ApartmentsHandler(link string) {
+func (h *TrackingHandler) ApartmentsHandler(link string, eventName string) {
 	log.Println("Processing...")
 	// Get all appartments from site
 
@@ -22,7 +22,7 @@ func (h *TrackingHandler) ApartmentsHandler(link string) {
 	// and if appartment was created - notify all bot subscribers
 	for _, outApp := range *outerApps {
 		eventType := h.GetEventTypeByApp(&outApp)
-		created, err := h.appService.CreateIfNotExist(&outApp)
+		created, err := h.appService.CreateIfNotExist(&outApp, eventType)
 		if err != nil {
 			fmt.Println(err.Error())
 		}
@@ -30,14 +30,14 @@ func (h *TrackingHandler) ApartmentsHandler(link string) {
 			h.NotifyAllSubscribers(&outApp, AppAdded, eventType)
 			continue
 		}
-		inApp, err := h.appService.GetById(outApp.ID)
+		inApp, err := h.appService.GetById(outApp.ID, eventType)
 		if err != nil {
 			fmt.Println(err.Error())
 		}
 		if inApp.Requested == outApp.Requested {
 			continue
 		}
-		_, err = h.appService.Update(&outApp, inApp.ID)
+		_, err = h.appService.Update(&outApp, inApp.ID, eventType)
 		if err != nil {
 			fmt.Println(err.Error())
 		}
@@ -50,7 +50,7 @@ func (h *TrackingHandler) ApartmentsHandler(link string) {
 
 	// Remove appartment from database if it was deleted from site
 	// and notify all bot subscribers
-	removedApps, err := h.appService.RemoveDeletedApps(outerApps)
+	removedApps, err := h.appService.RemoveDeletedApps(outerApps, eventName)
 	if err != nil {
 		fmt.Println(err.Error())
 	}
@@ -60,5 +60,5 @@ func (h *TrackingHandler) ApartmentsHandler(link string) {
 	}
 
 	log.Println("Sleep for 10 minutes...")
-	time.Sleep(time.Minute * 10)
+	time.Sleep(time.Second * 20)
 }
